:recycle: refactor: refactor code in ResultContainer
