+++ src/components/result/ResultContainer.tsx
@@ -1,20 +1,15 @@
 import axios from 'axios';
 import { useEffect, useState } from 'react';
-import {
-  Container,
-  RankWrapper,
-  Result,
-  ResultBox,
-  Text,
-  Title,
-} from './ResultPresenter';
+
+import RenderResult from './RenderResult';
+import { Container, ResultBox, Title } from './ResultPresenter';
 
 const ResultContainer = () => {
-  const [FEResult, setFEResult] = useState([]);
-  const [BEResult, setBEResult] = useState([]);
+  const [frontendResult, setFrontendResult] = useState([]);
+  const [backendResult, setBackendResult] = useState([]);
 
   useEffect(() => {
-    const fetchFEResult = async () => {
+    const fetchVoteResult = async () => {
       const FEResponse = await axios.get(
         'https://vote-mailedit.kro.kr/api/candidate/result?part=frontend'
       );
@@ -23,39 +18,19 @@ const ResultContainer = () => {
         'https://vote-mailedit.kro.kr/api/candidate/result?part=backend'
       );
 
-      setFEResult(FEResponse.data);
-      setBEResult(BEResponse.data);
+      setFrontendResult(FEResponse.data);
+      setBackendResult(BEResponse.data);
     };
 
-    fetchFEResult();
+    fetchVoteResult();
   }, []);
-
-  const rank = ['🥇', '🥈', '🥉', '4', '5', '6', '7', '8', '9', '10'];
 
   return (
     <Container>
-      <Title>Result</Title>
+      <Title>투표 결과</Title>
       <ResultBox>
-        <Result>
-          {FEResult.map((candidate: any, i: number) => (
-            <RankWrapper key={candidate.id}>
-              <Text>{rank[i]}</Text>
-              <Text>{candidate.name}</Text>
-
-              <Text>{candidate.vote_count}</Text>
-            </RankWrapper>
-          ))}
-        </Result>
-        <Result>
-          {BEResult.map((candidate: any, i: number) => (
-            <RankWrapper key={candidate.id}>
-              <Text>{rank[i]}</Text>
-              <Text>{candidate.name}</Text>
-
-              <Text>{candidate.vote_count}</Text>
-            </RankWrapper>
-          ))}
-        </Result>
+        <RenderResult resultList={frontendResult} />
+        <RenderResult resultList={backendResult} />
       </ResultBox>
     </Container>
   );
